Updated 3.3.4
Avatar Designer Window Adjustments - When editing the avatar, the window was wrapping under the bottom menu and the contents of the window did not fill the lower part of the window. Adjusted the vertical size of the window and set the background to black for the designer window only. Set the background opacity back to .2 when it closes and reopens.

Change Avatar Colors from RGB to Hex - Changed the avatar color database fields to use Hex values instead of RGB saved in the database. Everything now supports diffuse, specular, emissive, and ambient except for the user interface, which will be updated soon. This affected the JSON fetched data and save procedures, local and global avatars, and 3D Internet plugin in development. 

Added Avatar List to Admin Menu - Added an Admin Mode menu item for Avatars with a submenu item for Avatars List. In the near future, this will be used for admin to set the default Avatars, colors, and animations. 

Added Avatar Groups to Admin Menu - Added an Admin Mode submenu item for Avatar Groups. In the near future, this will be used for admin to group avatars and assign them to 3D Community Scenes.