Updated 3.3.5
Return Visit Avatar Loading - When you select an avatar for a 3D Website, it saves the useravatarid as a cookie. When you reload the page or return to it, the cookie is read and if it exists, loads the same avatar again. The global avatars had the right conditions, but the local avatars needed an update on the condition to make it work.

Last Animation Key Undefined Skeleton - Part of loading the animations to an avatar checks the last animation key so that it can append the next animation range to the avatar. When the skeleton does not have a previous animation loaded, this range or the skeleton may not be loaded yet and was causing an undefined error. Added the condition to check the skeleton existence before getting the key.

Load Avatar Animations requires Skeleton - When loading an avatar animation, if the animation to copy is not correct, it was breaking the process to load the remainder of the animations. Added code to check the skeleton before adding the animation.