Updated 3.3.5
Web Name Check Added - The webnamecheck.php page has been added to the connect folder. This will allow WordPress sites to confirm the availability of a website path for a community and building name.

WordPress Connect Page Added - In the connect folder, the wordpress.php webpage has been added. This is part of the WalkTheWeb WordPress Plugin features to be able to create a 3D Shopping Website in less than 5 minutes from a WooCommerce Online Store. It authenticates against the WalkTheWeb Global Login and requires local Admin or Host role access to interact with your WalkTheWeb instance.