Updated 3.3.5
Eliminated the use of Username - Many iterations ago, the primary key of userid was implemented for users mainly for the multiplayer system. Prior to that there was a username primary key. Now with default display name on the user account and each avatar having a separate display name, there is no longer a need for a username. This extends to all site code. Some instances when applicable, have been changed to use displayname while others were completely eliminated. The login process replaced username with email.