Updated 3.3.5
Renamed accesstoken to usertoken - Renamed accesstoken to usertoken in the users table and code. This usertoken is designed to authenticate from the WalkTheWeb global login. It has no connection to the local login.

Added new Host User Role - The Host user role will be used for Admin Mode permissions but limited to one or more Hosted 3D Websites. It will resemble an Admin for given 3D Communities, 3D Buildings, and 3D Things without being a full Admin on the server.

Added getFiles Function to Connect Class - The getFiles function adds the possibility to upload a file during a POST. The getFiles function was moved to the wtwdb class then extended to work from the handlers and connect classes.

Added Post Form Functions to Connect Class - The getPost function was moved to the wtwdb class then extended to work from the handlers and connect classes. The advantage is that the connect class can now support form POST data in addition to GET querystring based requests.

Removed username Database Fields - With the replacement of email over username for login, we have now removed username field from the users and userauthorizations tables. After the code update completes, you can remove these fields manually.

Removed addMustHave Function - During a new install many iterations ago, the process was downloading a number of Must Have 3D Things to get you started. It has been depreciated for some time, including the use of a non existent stored procedure. The Function and code that calls it has now been removed.

isUserNameAvailable Function Removed - The isUserNameAvailable function in the wtwusers PHP class has been removed since the username is no longer in use and does not need to be checked for uniqueness.

Changed getUsernameByEmail to getUserIdByEmail - In the wtwusers PHP Class, the function to verify if an email exists was returning the username and has been updated to return the userid, also renaming the function to match.

Eliminated the use of Username - Many iterations ago, the primary key of userid was implemented for users mainly for the multiplayer system. Prior to that there was a username primary key. Now with default display name on the user account and each avatar having a separate display name, there is no longer a need for a username. This extends to all site code. Some instances when applicable, have been changed to use displayname while others were completely eliminated. The login process replaced username with email.

Removed check Key for username and usernameoremail - In the checkKey function on the wtw_inputs.js, the username and usernameoremail valid type were removed since we no longer use username. The entries were updated to use email for the login.

Changed check Key for displayoremail - In the checkKey function on the wtw_inputs.js, the displayoremail valid type was changed to idoremail since we no longer use username. The entries were updated to use email or userid for the permissions granting.

Changed check Key for displayname - In the checkKey function on the wtw_inputs.js, the displayname valid type was updated to include the use of capital letters, lowercase, numbers, and very few symbols (hyphen, underscore, period, space, comma, and single quote).

Added Global Variable for Website URL - While working in Admin Mode, admin.php page, the web alias is not used by the site while in the browsing window. The added wtw_websiteurl can be used in JavaScript and $wtw->websiteurl can be used on the main website php pages.