Updated 3.4.0
UserToken Authentication for Downloads - The downloads process to get 3D Communities, 3D Buildings, and 3D Things uses the logged on session userid to create the new copy. When triggering a push based download like from WordPress WalkTheWeb Plugin, we needed a different way to authenticate the download, therefore used the WalkTheWeb Global Login - logged in usertoken. Each global user has a unique usertoken per website they connect, which keeps this process secure.

DownloadWeb Items Userid Reference - Some of the user id values inserted into the tables were using $handlers->userid while others used the assigned local variable $zuserid. With the addition of usertoken access, we needed to use the local variable for consistency.

SameSite Cookie Header - The SameSite Header is designed to limit or allow cookies to be passed to other 3rd party sites opened with the page. We added headers to all of the pertinent pages to limit or allow as necessary. This includes our main pages, handlers, and connect pages. 

Global User Avatar ID cleanup - The Global User Avatar Id was referenced by serveravatarid and globalavatarid in various parts of the code. We replaced the references with globaluseravatarid because it is the global WalkTheWeb version of useravatarid. This also allows us to reassign globalavatarid as a reference to a global WalkTheWeb avatars definition version of avatarid.

Global vs Server naming - There were a number of instances on the 3dnet.walktheweb.com site that set or sent fields referred to with names using server. Since each WalkTheWeb instance uses a serverinstanceid, we changed the hub key field references to names using the preface global.