Updated 3.4.0
Fixed Meta Tags - The meta tags for the website set the Title, Description, Image, and URL for the website.  A faulty SQL statement was stopping it from pulling the data from the tables. It has been corrected by fixing the statement and adding a communitydescription, buildingdescription, and thingdescription fields to the communities, buildings, and things tables respectively. The Description is also added to the Information pages under the Name.

Increased Default Walking Speed - Increased the default walking speed and walking animation speed to 2. It has a range of .1 to 2.9 in increments of .1 for both walking speed and walking animation. The default was 1, which makes this a third faster.