Updated 3.4.0
Email Server Functionality - We implemented the open source SwiftMail server as an Email Sending Service. This allows WalkTheWeb to connect to an SMTP service for sending emails. It includes Host, Port, Encryption, authentication with username and password, and settings for the from Email and Name. There is also a test email feature to make sure the settings are working.

Email Letter Case in Users Class - The users class was comparing emails and did not account for upper, lower, or mixed case letters. We changed the sql to compare using the term like to account for any servers that are case sensitive.

emailconfirm Field on users Table - Changed the emailconfirm field length to 255 on the users table to support larger hash values and to match the recoverpassword field length.

Save Settings without Callback - When saving settings using the JavaScript Function in wtw_common, it required a callback function. Now it allows a null or undefined callback function to just save the settings without the need of a response.

Send Email Function - We added a sendMail functions in JavaScript and PHP that allows you to set the sendto, cc, and bcc arrays of email addresses, subject, HTML message, and plain text message to generate a system sent email alert.