Updated 3.4.0
UserToken Authentication for Downloads - The downloads process to get 3D Communities, 3D Buildings, and 3D Things uses the logged on session userid to create the new copy. When triggering a push based download like from WordPress WalkTheWeb Plugin, we needed a different way to authenticate the download, therefore used the WalkTheWeb Global Login - logged in usertoken. Each global user has a unique usertoken per website they connect, which keeps this process secure.

Fixed Meta Tags - The meta tags for the website set the Title, Description, Image, and URL for the website.  A faulty SQL statement was stopping it from pulling the data from the tables. It has been corrected by fixing the statement and adding a communitydescription, buildingdescription, and thingdescription fields to the communities, buildings, and things tables respectively. The Description is also added to the Information pages under the Name.

SameSite Cookie Header - The SameSite Header is designed to limit or allow cookies to be passed to other 3rd party sites opened with the page. We added headers to all of the pertinent pages to limit or allow as necessary. This includes our main pages, handlers, and connect pages. 

Email Server Functionality - We implemented the open source SwiftMail server as an Email Sending Service. This allows WalkTheWeb to connect to an SMTP service for sending emails. It includes Host, Port, Encryption, authentication with username and password, and settings for the from Email and Name. There is also a test email feature to make sure the settings are working.

Global User Avatar ID cleanup - The Global User Avatar Id was referenced by serveravatarid and globalavatarid in various parts of the code. We replaced the references with globaluseravatarid because it is the global WalkTheWeb version of useravatarid. This also allows us to reassign globalavatarid as a reference to a global WalkTheWeb avatars definition version of avatarid.

Added Enable User Email Validation - We added a global variable WTW.enableEmailValidation with a 0 or 1 value to determine if the system will try to send and require email validation on login and create user accounts.

Send Email Function - We added a sendMail functions in JavaScript and PHP that allows you to set the sendto, cc, and bcc arrays of email addresses, subject, HTML message, and plain text message to generate a system sent email alert.