Updated 3.4.0
Increased Default Walking Speed - Increased the default walking speed and walking animation speed to 2. It has a range of .1 to 2.9 in increments of .1 for both walking speed and walking animation. The default was 1, which makes this a third faster.

Added Missing Logo Image - Added the missing image that shows in the meta preview for the website if a Preview Image is not found. The preview image can be set in the 3D Community -> Options and Settings -> 3D Community Snapshot. 

Fixed Meta Tags - The meta tags for the website set the Title, Description, Image, and URL for the website.  A faulty SQL statement was stopping it from pulling the data from the tables. It has been corrected by fixing the statement and adding a communitydescription, buildingdescription, and thingdescription fields to the communities, buildings, and things tables respectively. The Description is also added to the Information pages under the Name.

Colors for Default Avatars - The default avatars were not loading the colors with the new instances. Modified the useravatars.php connect page to include the colors based off the default avatarcolors table.

Email Server Functionality - We implemented the open source SwiftMail server as an Email Sending Service. This allows WalkTheWeb to connect to an SMTP service for sending emails. It includes Host, Port, Encryption, authentication with username and password, and settings for the from Email and Name. There is also a test email feature to make sure the settings are working.

Server Instance Variable - The serverinstanceid was sometimes referred to as serverinstance. For consistency and to match the database usage of it, it has now been set to serverinstanceid in all occurrences.

On Avatar Select return value - For use during plugins, the pluginsOnMyAvatarSelect function now returns a true or false as zloading to determine if the plugin is responsible for loading an avatar opposed to letting the default process load the avatar. This is a key in using Global Avatars.

Global User Avatar ID cleanup - The Global User Avatar Id was referenced by serveravatarid and globalavatarid in various parts of the code. We replaced the references with globaluseravatarid because it is the global WalkTheWeb version of useravatarid. This also allows us to reassign globalavatarid as a reference to a global WalkTheWeb avatars definition version of avatarid.

Global vs Server naming - There were a number of instances on the 3dnet.walktheweb.com site that set or sent fields referred to with names using server. Since each WalkTheWeb instance uses a serverinstanceid, we changed the hub key field references to names using the preface global.

Admin Mouse Overs - When user does not have certain permissions, they can still view a minimum admin menu. There were some errors caused by items not showing due to permissions. These items now check the availability before trying to set the values.

Save Settings without Callback - When saving settings using the JavaScript Function in wtw_common, it required a callback function. Now it allows a null or undefined callback function to just save the settings without the need of a response.

Added Enable User Email Validation - We added a global variable WTW.enableEmailValidation with a 0 or 1 value to determine if the system will try to send and require email validation on login and create user accounts.

Send Email Function - We added a sendMail functions in JavaScript and PHP that allows you to set the sendto, cc, and bcc arrays of email addresses, subject, HTML message, and plain text message to generate a system sent email alert. 

Created WalkTheWeb Update for Bot - Created a new WalkTheWeb Hub page that will post the WalkTheWeb updates to the database and Discord Bot! (Not part of the distribution, but interacts with the distribution)