Updated 3.4.0
On Avatar Select return value - For use during plugins, the pluginsOnMyAvatarSelect function now returns a true or false as zloading to determine if the plugin is responsible for loading an avatar opposed to letting the default process load the avatar. This is a key in using Global Avatars.

Global User Avatar ID cleanup - The Global User Avatar Id was referenced by serveravatarid and globalavatarid in various parts of the code. We replaced the references with globaluseravatarid because it is the global WalkTheWeb version of useravatarid. This also allows us to reassign globalavatarid as a reference to a global WalkTheWeb avatars definition version of avatarid.

Global vs Server naming - There were a number of instances on the 3dnet.walktheweb.com site that set or sent fields referred to with names using server. Since each WalkTheWeb instance uses a serverinstanceid, we changed the hub key field references to names using the preface global.