Updated 3.4.0
Email Server Functionality - We implemented the open source SwiftMail server as an Email Sending Service. This allows WalkTheWeb to connect to an SMTP service for sending emails. It includes Host, Port, Encryption, authentication with username and password, and settings for the from Email and Name. There is also a test email feature to make sure the settings are working. See SwiftMailer License and github.