Updated 3.4.0
Archive WalkTheWeb Update Details - You can now see the details of every WalkTheWeb update release and prerelease notes for the upcoming update on the Admin Menu - Dashboard - Updates page in the Admin of your WalkTheWeb 3D Website. It uses an accordion display, just click on the update to expand or collapse.