Updated 3.4.0
Multiplayer not Loading properly or Moving - The 3D Internet plugin was experiencing an issue since we implemented the server check. First of all, the server check required the server ip to be passed to several processes using move.js functions, and second, the movedata was incorrectly reading the server id for approval. All is corrected and working again!

3D Internet Plugin - Multiplayer and Chat - We are releasing the 3D Internet Plugin! This plugin provides Global WalkTheWeb Logins and Avatars to your WalkTheWeb instances for free. It also provides a connection to the main WalkTheWeb Multiplayer and Chat as an optional service for your WalkTheWeb instances! With a simple activation, you can now use our servers as your multiplayer and Chat servers for your WalkTheWeb Instances!!! Service Plans starting at only $120 per year (Just $10 USD per month).