Updated 3.4.1
Quick Fix, The Communities, Buildings, and Things tables each added a field for description. The fields did not make it into the last update. This updates the table definitions to force the update. You can manually add the fields: communitydescription (to table communities as varchar(255) default ''), buildingdescription (to table buildings as varchar(255) default '', and thingdescription (to table things as varchar(255) default '').