Updated 3.4.1
Update WalkTheWeb Process File Permissions - When the WalkTheWeb update process is called, it downloads the update then unzips it and replaces some of the files for the site. The unzip process was having a permissions issue so we added a permissions setting call before the unzip command to the newly downloaded file. This will correct the update unzip permissions.

Server Config Admin Page - Added a Server Settings Config page that allows the admin to easily edit the config file settings. There are a couple settings that cannot be changed on this form such as wtw_serverinstanceid, wtw_tableprefix, and wtw_contenturl. We also added a number of settings for wtw_adminname, wtw_chmod, wtw_umask, and for the upcoming FTP support, wtw_ftpuser, wtw_ftppassword, and wtw_ftpbase. Note that the passwords are stored in the config file using base64 encoding and this new form will set that automatically for you.

Send Test Email - The send test email on the wtw-swiftmailer plugin was set to convert the email string to an array, but it was already being converted in the handler. Removed the array function from the $zsendto variable and now the test email sends properly.

New Function in Uploads Class - We created a new function in the uploads class called getFilefromURL. We will be focusing all of the file transfers through this function so that we can append it with various methods as needed, all in one place. Currently it uses fopen and curl commands. FTP support will be added in the near future..

Config Variables and Field Names - The server settings in the config file are set from text input fields that did not match the names. Corrected the names, set passwords to use base64 encoding, renamed wtw_defaultfromemail to wtw_adminemail, and added additional support for the following fields: adminname, ftphost, ftpuser, ftppassword, ftpbase, umask (folder permissions), and chmod (file permissions). 

Config File Permissions - Using a defined constant in the config, we set up variables wtw_chmod and wtw_umask to easily allow the admin to configure the file permissions that are automatically added to uploaded files. wtw_chmod defaults to 755 and wtw_umask defaults to 0027.

Base64 Encode Password Constants - In order to base64 encode the database password in the config file, we set it up to automatically encode it if it is not. Then on every database function that calls the connection string (minimum occurrences since we use query functions), we added base64_decode function to the wtw_dbpassword constant. New instances of WalkTheWeb will automatically encode the database password on install.

Added description Fields to tables - The Communities, Buildings, and Things tables each added a field for description. The fields did not make it into the last update. This updates the table definitions to force the update. You can manually add the fields: communitydescription (to table communities as varchar(255) default ''), buildingdescription (to table buildings as varchar(255) default '', and thingdescription (to table things as varchar(255) default '').