Updated 3.4.1
Config Variables and Field Names - The server settings in the config file are set from text input fields that did not match the names. Corrected the names, set passwords to use base64 encoding, renamed wtw_defaultfromemail to wtw_adminemail, and added additional support for the following fields: adminname, ftphost, ftpuser, ftppassword, ftpbase, umask (folder permissions), and chmod (file permissions). 

Base64 Encode Password Constants - In order to base64 encode the database password in the config file, we set it up to automatically encode it if it is not. Then on every database function that calls the connection string (minimum occurrences since we use query functions), we added base64_decode function to the wtw_dbpassword constant. New instances of WalkTheWeb will automatically encode the database password on install.

Server Config Admin Page - Added a Server Settings Config page that allows the admin to easily edit the config file settings. There are a couple settings that cannot be changed on this form such as wtw_serverinstanceid, wtw_tableprefix, and wtw_contenturl. We also added a number of settings for wtw_adminname, wtw_chmod, wtw_umask, and for the upcoming FTP support, wtw_ftpuser, wtw_ftppassword, and wtw_ftpbase. Note that the passwords are stored in the config file using base64 encoding and this new form will set that automatically for you.