Updated 3.4.1
Added Cookies support to Connect and Handler Functions - To support cookies in php, the session_set_cookie_params have to be set before the session_start function. The main pages already have this in place, but the connect and handler class pages did not. We added the necessary code so now any cookies can be read from within the php code on these types of pages.