Update 3.4.1
Corrected Animation Restarts in Extreme Zone - When an animation is part of a mold (mesh) and is assigned to load in the Extreme Load Zone, If the avatar walked out of the extreme zone and back in, the animation would not restart. Basically, the animation from the first loading was not completely disposing, causing it to think it was already loaded and running. The order needed to be modified to remove the animation completely on dispose of the mold, so that it could fully reload on next attempt to enter the extreme load zone (Previously, it was only working in other load zones). We moved the disposeMoldEvent function call to sooner in the disposeClean function and removed the constraint so that it happens on all child items of a disposed object and now it works correctly.