Updated 3.4.1
Corrected Sound cut-off on Unload - The Sound was not stopping when it was attached to a mold in the extreme load zone. Then when you would reenter it, the sound would over-lap the last sound and play again. The solution was similar to the Animation restarting issue, we moved the disposeSoundAndLights function to happen sooner in the disposeClean function. This corrected the sound issue.