Updated 3.4.1
Added Fields to Communities Table - We added additional fields for setting the First 3D Building into a 3D Community. The First Building location is used by the automated processes to add a 3D Building (example 3D Store) to a 3D Community with minimal input. We already had buildingpositionx, buildingpositiony, buildingpositionz, and buildingrotationz.We now added buildingscalingx, buildingscalingy, buildingscalingz, buildingrotationx, and buildingrotationy to match the available settings for a typical connecting grid. 

Set First 3D Building Location on 3D Community - In order to make new 3D Community Scenes work well in automated processes, like quick start 3D Shopping Stores, we needed to have a method to select the location of the first 3D Building that will be added to the 3D Community. These settings involve position, scaling, and rotation and become the connecting grid settings for that first 3D Building in the 3D Community. The values are stored in the communities table. When setting the value, there is a place marker that can be moved to visualize the location. It represents about a 120x120 unit foundation, so make sure to leave room on all sides for larger buildings as desired.  The admin menu item to set the First 3D Building location is named Set First 3D Building and has been added to the Options and Settings menu of 3D Communities.

Added First Building Location to Copy and Share - The new fields for the First 3D Building in the communities table have been added to the processes for Share 3D Communities, Download Shared 3D Communities, and Copy 3D Communities. In addition, the fields have been added to all relevant connect, handler, and function files.