Updated 3.4.1
Corrected Animation Restarts in Extreme Zone - When an animation is part of a mold (mesh) and is assigned to load in the Extreme Load Zone, If the avatar walked out of the extreme zone and back in, the animation would not restart. Basically, the animation from the first loading was not completely disposing, causing it to think it was already loaded and running. The order needed to be modified to remove the animation completely on dispose of the mold, so that it could fully reload on next attempt to enter the extreme load zone (Previously, it was only working in other load zones). We moved the disposeMoldEvent function call to sooner in the disposeClean function and removed the constraint so that it happens on all child items of a disposed object and now it works correctly.

Corrected Sound cut-off on Unload - The Sound was not stopping when it was attached to a mold in the extreme load zone. Then when you would reenter it, the sound would over-lap the last sound and play again. The solution was similar to the Animation restarting issue, we moved the disposeSoundAndLights function to happen sooner in the disposeClean function. This corrected the sound issue.

Added Fields to Communities Table - We added additional fields for setting the First 3D Building into a 3D Community. The First Building location is used by the automated processes to add a 3D Building (example 3D Store) to a 3D Community with minimal input. We already had buildingpositionx, buildingpositiony, buildingpositionz, and buildingrotationz.We now added buildingscalingx, buildingscalingy, buildingscalingz, buildingrotationx, and buildingrotationy to match the available settings for a typical connecting grid. 

Set First 3D Building Location on 3D Community - In order to make new 3D Community Scenes work well in automated processes, like quick start 3D Shopping Stores, we needed to have a method to select the location of the first 3D Building that will be added to the 3D Community. These settings involve position, scaling, and rotation and become the connecting grid settings for that first 3D Building in the 3D Community. The values are stored in the communities table. When setting the value, there is a place marker that can be moved to visualize the location. It represents about a 120x120 unit foundation, so make sure to leave room on all sides for larger buildings as desired.  The admin menu item to set the First 3D Building location is named Set First 3D Building and has been added to the Options and Settings menu of 3D Communities.

Added First Building Location to Copy and Share - The new fields for the First 3D Building in the communities table have been added to the processes for Share 3D Communities, Download Shared 3D Communities, and Copy 3D Communities. In addition, the fields have been added to all relevant connect, handler, and function files.