Updated 3.4.1
Set First 3D Building Location on 3D Community - In order to make new 3D Community Scenes work well in automated processes, like quick start 3D Shopping Stores, we needed to have a method to select the location of the first 3D Building that will be added to the 3D Community. These settings involve position, scaling, and rotation and become the connecting grid settings for that first 3D Building in the 3D Community. The values are stored in the communities table. When setting the value, there is a place marker that can be moved to visualize the location. It represents about a 120x120 unit foundation, so make sure to leave room on all sides for larger buildings as desired.  The admin menu item to set the First 3D Building location is named Set First 3D Building and has been added to the Options and Settings menu of 3D Communities.

These files are the 3D Model - 3D Building Marker