Updated 3.4.1
Added Description to Share Process - In addition to the share description, communities, buildings, and things tables added a description field that is used by the meta tags when browsing the 3D Website. For reference, we prevously added the communitydescription to the communities table, buildingdescription to the buildings table, and thingdescription to the things table. Now we included these description fields when sharing or getting a shared 3D Web from downloads.