Updated 3.4.1
Added Fields to 3D Shopping Plugin Tables - We have added 2 new fields to the stores table for 3D Shopping plugin: woocommercekeynew and woocommercesecretnew. They will assist with a process to push an update to the WooCommerce key and secret from the soon to be released updated WordPress - WalkTheWeb 3D Shopping plugin.