Updated 3.4.1
Repurposed and Renamed Table UserAPIKeys - The table userapikeys was no longer in use so it has been repurposed for App level API Keys tracking. The table was renamed to apikeys and fields removed the term user, and fields were added for appid allowing 32 characters, appname 255 characters, approved flag as int, approveddate, and approveduserid. This will be used to allow outside apps with proper permissions to perform select secure tasks within the WalkTheWeb instance.

Added decode64 PHP Function - There are a number of times where we expect base64_encoded string, but if it is not formatted correctly, it can cause an error that stops the code from completing. We created a function in the wtwdb class and extended it to the wtwconnect and wtwhandlers classes so that you can easily pass in the text and get either the base64_decode text or blank if it is not base64. This way we can still gracefully process the remaining code and just check for empty text.

Replaced base64_decode with new decode64 Function - The base64_decode PHP function was generally used after data transfers in JSON and form posts which are generally in the Connect, Handlers, and Functions. We have now updated them to use the new decode64 function; example $decodedtext = $wtwconnect->decide64(encoded64text); This will now return a blank value instead of causing a halting error in code. 

New API Access Settings - To achieve a high level of connectivity, there are times when we need an outside application to access or set data within a WalkTheWeb instance. In Admin, under the Settings menu, we have added an API Keys Access menu option. This takes you to a page that can create, edit, approve, deny, reassign, and delete API Keys.

New API Handler, Class, and Connect Files - All of the functions for the new Admin API Keys Access functionality are grouped into a new class called wtwapi. This class can be used from other classes, connect files, and handlers. The functions include getAPIKeys, getAPIKey, approveAPIKey, saveAPIKey, and deleteAPIKey. Each of these functions require Admin access to the WalkTheWeb instance.

Removed old wtw-multiplayer Code - The JavaScript getJSON function in utilities had legacy code to support the old wtw-multiplayer plugin. This plugin has been replaced by the new wtw-3dinternet plugin. The legacy code has been removed.