Updated 3.4.1.
Added 3D Object Filter in Media Library - In Admin under the the media Library, 3D Objects page, when the list becomes long, it is more difficult to look through it and search for a particular 3D Object. We added a text box for Name Filter. When you type in this box, the 3D Objects listed will filter to only show the ones with matching file Names. This only filters by the main file name that the 3D Object is labeled as shown on this page. Clearing the box will restore all of the 3D Objects.