Updated 3.4.1
Added Submenu to Media Library - For consistency and quick access, we added a submenu under the Media Library in the Admin Menu. The submenu includes Files, 3D Objects, and WalkTheWeb Downloads.

Renamed 3D Objects to 3D Models - In the Media Library, we renamed 3D Objects to 3D Models. This name is more accurate for what is being uploaded. This not only includes the labels on the menu and buttons, but the related HTML IDs have also been renamed. This will help avoid confusion since 3D Objects is a much broader category of items in a 3D Scene. 3D Models are imported 3D Art usually created by a program like Blender.