Updated 3.4.1
Optional Parameters for DownloadWeb Function - In the wtwcommunities php Class, the downloadWeb function now has 16 parameters since we added additional scaling and rotation options. The instances on previous updates only had 11 parameters and it now uses 16. We changed the parameters to make the position, scaling, and rotation optional by assigning a default value. These parameters are used for the automation processes that create a 3D Community with a preselected 3D Building as the first building in the 3D Scene. This allows the function to now be called from older or newer instances of WalkTheWeb.