Updated 3.4.1
File Upload Path for Users Update - The 3D Models uploads requires a temp folder under the user folder called media. The folder was not being created. Added the code to create and set the permissions for the media folder in the uploads class checkContentFolders function. 

Optional Parameters for DownloadWeb Function - In the wtwcommunities php Class, the downloadWeb function now has 16 parameters since we added additional scaling and rotation options. The instances on previous updates only had 11 parameters and it now uses 16. We changed the parameters to make the position, scaling, and rotation optional by assigning a default value. These parameters are used for the automation processes that create a 3D Community with a preselected 3D Building as the first building in the 3D Scene. This allows the function to now be called from older or newer instances of WalkTheWeb.

Renamed 3D Objects to 3D Models - In the Media Library, we renamed 3D Objects to 3D Models. This name is more accurate for what is being uploaded. This not only includes the labels on the menu and buttons, but the related HTML IDs have also been renamed. This will help avoid confusion since 3D Objects is a much broader category of items in a 3D Scene. 3D Models are imported 3D Art usually created by a program like Blender.

Added Submenu to Media Library - For consistency and quick access, we added a submenu under the Media Library in the Admin Menu. The submenu includes Files, 3D Objects, and WalkTheWeb Downloads.

Added 3D Object Filter in Media Library - In Admin under the the media Library, 3D Objects page, when the list becomes long, it is more difficult to look through it and search for a particular 3D Object. We added a text box for Name Filter. When you type in this box, the 3D Objects listed will filter to only show the ones with matching file Names. This only filters by the main file name that the 3D Object is labeled as shown on this page. Clearing the box will restore all of the 3D Objects.