Updated 3.4.1
Corrected 3D Thing not showing when added - When you add a 3D Thing into a 3D Building or 3D Community or add a 3D Building into a 3D Community, the added 3D Web does not show until your avatar moves to trigger a movement in the Load Action Zone. We now added the setting after Action Zones are added to trigger if the avatar is inside the Action Zone. This applies to Load Zones as well as all others; for example, swinging door would open if the avatar is in the Action Zone.

First Building Setting Corrected - The First 3D Building setting is used on 3D Communities in Options and Settings to set the location of the first added 3D Building for the automated 3D Website creation apps. While it was saving the settings correctly, it was not reloading the past settings. We needed to change the connect page from communities.php to community.php and add the communities reference to the result set to read the values.

Rename JavaScript Functions loadBuildingConnectingGrids and loadBuildingThingsLoadZones - We renamed JavaScript Function loadBuildingConnectingGrids to loadChildConnectingGrids since it references 3D Buildings in 3D Communities and 3D Things in 3D Buildings or 3D Communities.  We also renamed loadBuildingThingsLoadZones to loadChildLoadZones for the same scope reason.

Fixed Avatar not Disposing Properly - Occasionally, most often in multiplayer mode, an avatar would get orphaned when the base is disposed of but the 3D Model had not finished loading. This was obvious since the model did not get scaled and would show up about 10 times the normal size. We now are checking to make sure the base still exists at the end of the asynchronously loaded 3D Model. Otherwise, all child parts are disposed of from the scene.

Dispose Clean additions and Changes - The disposeClean JavasScript function in the wtw_common.js file needed a couple of improvements. First, the plugin dispose clean was only activating if the base mold was found in the scene. We moved it before that section so that it executes regardless of the base mold existence. This helps with cleaning the child orphaned meshes which also have the same base name with an additional child mold name appended. Second, using the base name, we are now checking the scene meshes for child based names and dispose of them for babylonfiles (3D Models). Finally, the same addition to the 3D Shopping plugin disposeClean function was added to the Store based items, most of which have child meshes.

3D Molds and Disposing of Child Objects - 3D Molds are loaded in 2 parts, first the invisible box used as a parent for positioning, scaling, and rotation, then the rest of the 3D Mold is loaded using an asynchronous call which may take a little time to fetch and load. There are times when the base is created and before the rest of the 3D Mold is loaded, the avatar leaves the load zone and the base is deleted leaving the 3D Model orphaned in the center of the scene without a parent. We added code to the end of all of the ImportMeshAsync methods to check if the base is still in the scene and if not, dispose of the child 3D Model parts. 

3D Shopping Products Populated - When the 3D Product Displays are created, the first one triggers a call to the online store to get a products list. Before the products were loading asynchronously, the product info was retrieved, then all of the displays were updated using this information. Now the displays were created before the information was full retrieved and an additional function call had to be added to provide one last attempt to fill the product displays when completes retrieving the product information. Final results, it is working much faster and completing the process.

3D Shopping Plugin Corrected GetMolds - The wtw-shopping-getMolds.php connect file in the 3D Shopping Plugin needed updates to make it work from any webid. The select SQL aliased the communitymoldid, buildingmoldid, or thingmoldid as moldid and now we have corrected the name retrieved from the array to be moldid to match. In addition it was redundant to check the webid after each individual table already checked it so that part of the where clause has been removed. The connect file is working as expected now.